backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyswisseph
pytz
pydantic[email]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
httpx
